+++ components/zcash_protocol/src/local_consensus.rs
@@ -33,6 +33,7 @@ use crate::consensus::{BlockHeight, NetworkType, NetworkUpgrade, Parameters};
 ///         nu6_3: Some(BlockHeight::from_u32(1)),
 ///         #[cfg(zcash_unstable = "nu7")]
 ///         nu7: Some(BlockHeight::from_u32(1)),
+///         nu_tachyon: Some(BlockHeight::from_u32(1)),
 ///     };
 ///     ```
 ///
@@ -50,6 +51,7 @@ pub struct LocalNetwork {
     pub nu6_3: Option<BlockHeight>,
     #[cfg(zcash_unstable = "nu7")]
     pub nu7: Option<BlockHeight>,
+    pub nu_tachyon: Option<BlockHeight>,
 }
 
 /// Parameters implementation for `LocalNetwork`
@@ -72,6 +74,7 @@ impl Parameters for LocalNetwork {
             NetworkUpgrade::Nu6_3 => self.nu6_3,
             #[cfg(zcash_unstable = "nu7")]
             NetworkUpgrade::Nu7 => self.nu7,
+            NetworkUpgrade::NuTachyon => self.nu_tachyon,
         }
     }
 }
@@ -98,6 +101,7 @@ mod tests {
         let expected_nu6_3 = BlockHeight::from_u32(10);
         #[cfg(zcash_unstable = "nu7")]
         let expected_nu7 = BlockHeight::from_u32(11);
+        let expected_nu_tachyon = BlockHeight::from_u32(12);
 
         let regtest = LocalNetwork {
             overwinter: Some(expected_overwinter),
@@ -112,6 +116,7 @@ mod tests {
             nu6_3: Some(expected_nu6_3),
             #[cfg(zcash_unstable = "nu7")]
             nu7: Some(expected_nu7),
+            nu_tachyon: Some(expected_nu_tachyon),
         };
 
         assert!(regtest.is_nu_active(NetworkUpgrade::Overwinter, expected_overwinter));
@@ -127,6 +132,8 @@ mod tests {
         // nu7 must not be activated at or below the nu6_3 height
         #[cfg(zcash_unstable = "nu7")]
         assert!(!regtest.is_nu_active(NetworkUpgrade::Nu7, expected_nu6_3));
+        assert!(!regtest.is_nu_active(NetworkUpgrade::NuTachyon, expected_nu6_3));
+        assert!(regtest.is_nu_active(NetworkUpgrade::NuTachyon, expected_nu_tachyon));
     }
 
     #[test]
@@ -143,6 +150,7 @@ mod tests {
         let expected_nu6_3 = BlockHeight::from_u32(10);
         #[cfg(zcash_unstable = "nu7")]
         let expected_nu7 = BlockHeight::from_u32(11);
+        let expected_nu_tachyon = BlockHeight::from_u32(12);
 
         let regtest = LocalNetwork {
             overwinter: Some(expected_overwinter),
@@ -157,6 +165,7 @@ mod tests {
             nu6_3: Some(expected_nu6_3),
             #[cfg(zcash_unstable = "nu7")]
             nu7: Some(expected_nu7),
+            nu_tachyon: Some(expected_nu_tachyon),
         };
 
         assert_eq!(
@@ -204,6 +213,10 @@ mod tests {
             regtest.activation_height(NetworkUpgrade::Nu7),
             Some(expected_nu7)
         );
+        assert_eq!(
+            regtest.activation_height(NetworkUpgrade::NuTachyon),
+            Some(expected_nu_tachyon)
+        );
     }
 
     #[test]
@@ -220,6 +233,7 @@ mod tests {
         let expected_nu6_3 = BlockHeight::from_u32(10);
         #[cfg(zcash_unstable = "nu7")]
         let expected_nu7 = BlockHeight::from_u32(11);
+        let expected_nu_tachyon = BlockHeight::from_u32(12);
 
         let regtest = LocalNetwork {
             overwinter: Some(expected_overwinter),
@@ -234,6 +248,7 @@ mod tests {
             nu6_3: Some(expected_nu6_3),
             #[cfg(zcash_unstable = "nu7")]
             nu7: Some(expected_nu7),
+            nu_tachyon: Some(expected_nu_tachyon),
         };
 
         assert_eq!(regtest.coin_type(), constants::regtest::COIN_TYPE);