+++ pczt/tests/end_to_end.rs
@@ -1537,6 +1537,7 @@ fn pre_nu6_3_test_network() -> zcash_protocol::local_consensus::LocalNetwork {
         nu6_3: None,
         #[cfg(zcash_unstable = "nu7")]
         nu7: None,
+        nu_tachyon: None,
     }
 }
 
@@ -1555,6 +1556,7 @@ fn nu6_3_test_network() -> zcash_protocol::local_consensus::LocalNetwork {
         nu6_3: Some(BlockHeight::from_u32(10)),
         #[cfg(zcash_unstable = "nu7")]
         nu7: None,
+        nu_tachyon: None,
     }
 }